DOC: Clean termination when running as a daemon
Using the GPIO functions as a part of a long
running daemon seemed to be a common use case.

When doing so using the pigpio module, unless
we do a clean termination in the event of
service stop and restarts, we get errors in the
case of subsequent attempts at initializing the
pigpio library.

If termination is not done, problems similar to the following
may occur: https://goo.gl/gkYvch

Hence, added clean termination step in the Usage section.